Change mAmp to type float

+++ bms_software/Core/Src/bms_current_reader.c
@@ -12,9 +12,9 @@ uint16_t adc_resistor_drop(uint16_t adc_pre, uint16_t adc_post){
 	else return 0;
 }
 
-uint16_t convert_adc_to_mAmp(uint16_t adc_diff){
-	uint16_t voltage = (SYSTEM_VOLTAGE*adc_diff)/ADC_RESOLUTION;
-	uint16_t mAmp = (voltage / CHARGING_RESISTANCE) * MILLI;
+float convert_adc_to_mAmp(uint16_t adc_diff){
+	float voltage = (SYSTEM_VOLTAGE*adc_diff)/ADC_RESOLUTION;
+	float mAmp = (voltage / CHARGING_RESISTANCE) * MILLI;
 	return mAmp;
 }
